+++ pkg/client/client_darwin.go
@@ -12,7 +12,7 @@ func (c *Client) upgradeWindows(_ interface{}) {}
 
 func (c *Client) AutoWatchUpdate() {}
 
-func (c *Client) checkReloadSignal(sigc os.Signal) {
+func (c *Client) checkReloadSignal(sigc os.Signal) error {
 	if sigc == syscall.SIGUSR1 && c.Flags.ConfigFile != "" {
 		c.Printf("Writing Config File! Caught Signal: %v", sigc)
 
@@ -20,10 +20,13 @@ func (c *Client) checkReloadSignal(sigc os.Signal) {
 			c.Errorf("Writing Config File: %v", err)
 		}
 	} else {
-		c.reloadConfiguration("caught signal: " + sigc.String())
+		return c.reloadConfiguration("Caught Signal: " + sigc.String())
 	}
+
+	return nil
 }
 
-func (c *Client) setReloadSignals() {
+func (c *Client) setSignals() {
+	signal.Notify(c.sigkil, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT)
 	signal.Notify(c.sighup, syscall.SIGHUP, syscall.SIGUSR1)
 }
