+++ pkg/client/handlers.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/Notifiarr/notifiarr/pkg/bindata"
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
-	"github.com/Notifiarr/notifiarr/pkg/plex"
 	"github.com/Notifiarr/notifiarr/pkg/ui"
 	"github.com/gorilla/mux"
 	"golift.io/starr"
@@ -21,7 +20,7 @@ import (
 func (c *Client) internalHandlers() {
 	c.Config.HandleAPIpath("", "slow", c.slowResponse, "HEAD") // log testing
 	c.Config.HandleAPIpath("", "status", c.statusResponse, "GET", "HEAD")
-	c.Config.HandleAPIpath("", "version", c.versionResponse, "GET", "HEAD")
+	c.Config.HandleAPIpath("", "version", c.notifiarr.VersionHandler, "GET", "HEAD")
 	c.Config.HandleAPIpath("", "info", c.updateInfo, "PUT")
 	c.Config.HandleAPIpath("", "info/alert", c.updateInfoAlert, "PUT")
 	c.Config.HandleAPIpath("", "trigger/{trigger:[0-9a-z-]+}", c.handleTrigger, "GET")
@@ -33,13 +32,12 @@ func (c *Client) internalHandlers() {
 
 		tokens := fmt.Sprintf("{token:%s|%s}", c.Config.Plex.Token, c.Config.Apps.APIKey)
 		c.Config.Router.Handle("/plex",
-			http.HandlerFunc(c.plexIncoming)).Methods("POST").Queries("token", tokens)
+			http.HandlerFunc(c.notifiarr.PlexHandler)).Methods("POST").Queries("token", tokens)
 
 		if c.Config.URLBase != "/" {
 			// Allow plex to use the base url too.
-			// If this causes a panic, they probably set urlbase to "//" or something.
 			c.Config.Router.Handle(path.Join(c.Config.URLBase, "plex"),
-				http.HandlerFunc(c.plexIncoming)).Methods("POST").Queries("token", tokens)
+				http.HandlerFunc(c.notifiarr.PlexHandler)).Methods("POST").Queries("token", tokens)
 		}
 	}
 
@@ -98,81 +96,6 @@ func (c *Client) updateInfoAlert(r *http.Request) (int, interface{}) {
 	}()
 
 	return code, err
-}
-
-type appStatus struct {
-	Radarr  []*conTest `json:"radarr"`
-	Readarr []*conTest `json:"readarr"`
-	Sonarr  []*conTest `json:"sonarr"`
-	Lidarr  []*conTest `json:"lidarr"`
-	Plex    []*conTest `json:"plex"`
-}
-
-type conTest struct {
-	Instance int         `json:"instance"`
-	Up       bool        `json:"up"`
-	Status   interface{} `json:"systemStatus,omitempty"`
-}
-
-// versionResponse returns application run and build time data and application statuses: /api/version.
-func (c *Client) versionResponse(r *http.Request) (int, interface{}) {
-	var (
-		output = c.notify.Info()
-		rad    = make([]*conTest, len(c.Config.Radarr))
-		read   = make([]*conTest, len(c.Config.Readarr))
-		son    = make([]*conTest, len(c.Config.Sonarr))
-		lid    = make([]*conTest, len(c.Config.Lidarr))
-		status = &appStatus{Radarr: rad, Readarr: read, Sonarr: son, Lidarr: lid}
-	)
-
-	for i, app := range c.Config.Radarr {
-		stat, err := app.GetSystemStatus()
-		rad[i] = &conTest{Instance: i + 1, Up: err == nil, Status: stat}
-	}
-
-	for i, app := range c.Config.Readarr {
-		stat, err := app.GetSystemStatus()
-		read[i] = &conTest{Instance: i + 1, Up: err == nil, Status: stat}
-	}
-
-	for i, app := range c.Config.Sonarr {
-		stat, err := app.GetSystemStatus()
-		son[i] = &conTest{Instance: i + 1, Up: err == nil, Status: stat}
-	}
-
-	for i, app := range c.Config.Lidarr {
-		stat, err := app.GetSystemStatus()
-		lid[i] = &conTest{Instance: i + 1, Up: err == nil, Status: stat}
-	}
-
-	if c.Config.Plex.Configured() {
-		stat, err := c.Config.Plex.GetInfo()
-		if stat == nil {
-			stat = &plex.PMSInfo{}
-		}
-
-		status.Plex = []*conTest{{
-			Instance: 1,
-			Up:       err == nil,
-			Status: map[string]interface{}{
-				"friendlyName":             stat.FriendlyName,
-				"version":                  stat.Version,
-				"updatedAt":                stat.UpdatedAt,
-				"platform":                 stat.Platform,
-				"platformVersion":          stat.PlatformVersion,
-				"size":                     stat.Size,
-				"myPlexSigninState":        stat.MyPlexSigninState,
-				"myPlexSubscription":       stat.MyPlexSubscription,
-				"pushNotifications":        stat.PushNotifications,
-				"streamingBrainVersion":    stat.StreamingBrainVersion,
-				"streamingBrainABRVersion": stat.StreamingBrainABRVersion,
-			},
-		}}
-	}
-
-	output["app_status"] = status
-
-	return http.StatusOK, output
 }
 
 // notFound is the handler for paths that are not found: 404s.
@@ -251,7 +174,7 @@ func (c *Client) handleTrigger(r *http.Request) (int, interface{}) {
 
 	switch {
 	case trigger == "cfsync":
-		c.notify.SyncCF(false)
+		c.notifiarr.Trigger.SyncCF(false)
 	case trigger == "services" && c.Config.Services.Disabled:
 		return http.StatusNotImplemented, "services not enabled"
 	case trigger == "services":
@@ -259,13 +182,13 @@ func (c *Client) handleTrigger(r *http.Request) (int, interface{}) {
 	case trigger == "sessions" && !c.Config.Plex.Configured():
 		return http.StatusNotImplemented, "sessions not enabled"
 	case trigger == "sessions":
-		c.notify.SendPlexSessions(apiTrigger)
+		c.notifiarr.Trigger.SendPlexSessions(apiTrigger)
 	case trigger == "stuckitems":
-		c.notify.SendFinishedQueueItems(c.notify.BaseURL)
+		c.notifiarr.Trigger.SendFinishedQueueItems(c.notifiarr.BaseURL)
 	case trigger == "dashboard":
-		c.notify.GetState()
+		c.notifiarr.Trigger.GetState()
 	case trigger == "snapshot":
-		c.notify.SendSnapshot(apiTrigger)
+		c.notifiarr.Trigger.SendSnapshot(apiTrigger)
 	default:
 		return http.StatusBadRequest, "unknown trigger '" + trigger + "'"
 	}
