+++ pkg/client/start.go
@@ -6,17 +6,15 @@
 package client
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"net/http"
 	"os"
-	"os/signal"
-	"runtime"
-	"syscall"
 	"time"
 
 	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/configfile"
-	"github.com/Notifiarr/notifiarr/pkg/curl"
 	"github.com/Notifiarr/notifiarr/pkg/logs"
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
 	"github.com/Notifiarr/notifiarr/pkg/notifiarr"
@@ -26,24 +24,22 @@ import (
 	"github.com/Notifiarr/notifiarr/pkg/update"
 	flag "github.com/spf13/pflag"
 	"golift.io/cnfg"
-	"golift.io/rotatorr"
 	"golift.io/version"
 )
 
 // Client stores all the running data.
 type Client struct {
 	*logs.Logger
-	Flags  *configfile.Flags
-	Config *configfile.Config
-	server *http.Server
-	sigkil chan os.Signal
-	sighup chan os.Signal
-	menu   map[string]ui.MenuItem
-	info   string
-	notify *notifiarr.Config
-	alert  *logs.Cooler
-	plex   *logs.Timer
-	newCon bool
+	Flags     *configfile.Flags
+	Config    *configfile.Config
+	server    *http.Server
+	sigkil    chan os.Signal
+	sighup    chan os.Signal
+	menu      map[string]ui.MenuItem
+	info      string
+	notifiarr *notifiarr.Config
+	alert     *logs.Cooler
+	newCon    bool
 }
 
 // Errors returned by this package.
@@ -53,13 +49,14 @@ var (
 
 // NewDefaults returns a new Client pointer with default settings.
 func NewDefaults() *Client {
+	logger := logs.New() // This persists throughout the app.
+
 	return &Client{
 		sigkil: make(chan os.Signal, 1),
 		sighup: make(chan os.Signal, 1),
 		menu:   make(map[string]ui.MenuItem),
-		plex:   &logs.Timer{},
 		alert:  &logs.Cooler{},
-		Logger: logs.New(),
+		Logger: logger,
 		Config: &configfile.Config{
 			Apps: &apps.Apps{
 				URLBase: "/",
@@ -67,6 +64,7 @@ func NewDefaults() *Client {
 			Services: &services.Config{
 				Interval: cnfg.Duration{Duration: services.DefaultSendInterval},
 				Parallel: 1,
+				Logger:   logger,
 			},
 			BindAddr: mnd.DefaultBindAddr,
 			Snapshot: &snapshot.Config{
@@ -91,40 +89,38 @@ func Start() error {
 	c.Flags.ParseArgs(os.Args[1:])
 
 	switch {
-	case c.Flags.VerReq:
-		fmt.Println(version.Print(c.Flags.Name()))
-		return nil // print version and exit.
-	case c.Flags.PSlist:
+	case c.Flags.VerReq: // print version and exit.
+		fmt.Println(version.Print(c.Flags.Name())) //nolint:forbidigo
+		return nil
+	case c.Flags.PSlist: // print process list and exit.
 		return printProcessList()
-	case c.Flags.Curl != "":
-		resp, body, err := curl.Get(c.Flags.Curl) //nolint:bodyclose // it's already closed.
-		if err != nil {
-			return fmt.Errorf("getting URL '%s': %w", c.Flags.Curl, err)
-		}
+	case c.Flags.Curl != "": // curl a URL and exit.
+		return curlURL(c.Flags.Curl)
+	}
 
-		curl.Print(resp, body)
-
+	switch msg, err := c.loadConfiguration(); {
+	case err != nil:
+		return fmt.Errorf("%s: %w", msg, err)
+	case c.Flags.Restart:
 		return nil
+	case c.Flags.Write != "":
+		c.Printf("==> %s", msg)
+		return c.forceWriteWithExit(c.Flags.Write)
+	case c.Config.APIKey == "":
+		return fmt.Errorf("%s: %w %s_API_KEY", msg, ErrNilAPIKey, c.Flags.EnvPrefix)
+	default:
+		c.Logger.SetupLogging(c.Config.LogConfig)
+		c.Printf("%s v%s-%s Starting! [PID: %v] %v",
+			c.Flags.Name(), version.Version, version.Revision, os.Getpid(), time.Now())
+		c.Printf("==> %s", msg)
+		c.printUpdateMessage()
 	}
 
-	if err := c.config(); err != nil {
-		_, _ = ui.Error(mnd.Title, err.Error())
-		return err
-	} else if c.Flags.Restart || c.Flags.Write != "" {
-		return nil
-	}
-
-	if err := c.start(); err != nil {
-		_, _ = ui.Error(mnd.Title, err.Error())
-		return err
-	}
-
-	return nil
+	return c.start()
 }
 
-func (c *Client) config() error {
-	var msg string
-
+// loadConfiguration brings in, and sometimes creates, the initial running confguration.
+func (c *Client) loadConfiguration() (msg string, err error) {
 	// Find or write a config file. This does not parse it.
 	// A config file is only written when none is found on Windows, macOS (GUI App only), or Docker.
 	// And in the case of Docker, only if `/config` is a mounted volume.
@@ -132,144 +128,54 @@ func (c *Client) config() error {
 	c.Flags.ConfigFile, c.newCon, msg = c.Config.FindAndReturn(c.Flags.ConfigFile, write)
 
 	if c.Flags.Restart {
-		return update.Restart(&update.Command{ //nolint:wrapcheck
+		return msg, update.Restart(&update.Command{ //nolint:wrapcheck
 			Path: os.Args[0],
 			Args: []string{"--updated", "--config", c.Flags.ConfigFile},
 		})
 	}
 
 	// Parse the config file and environment variables.
-	if err := c.Config.Get(c.Flags.ConfigFile, c.Flags.EnvPrefix); err != nil {
-		return fmt.Errorf("%s: %w", msg, err)
+	c.notifiarr, err = c.Config.Get(c.Flags.ConfigFile, c.Flags.EnvPrefix, c.Logger)
+	if err != nil {
+		return msg, fmt.Errorf("getting config: %w", err)
 	}
 
-	// If c.Flags.write is set it will force-write the read-config to the provided file path.
-	if c.Flags.Write != "" {
-		return c.forceWriteWithExit(c.Flags.Write, msg)
-	}
-
-	c.startupMessage([]string{msg})
-
-	return nil
+	return msg, nil
 }
 
-func (c *Client) forceWriteWithExit(f, msg string) error {
-	if f == "-" {
-		f = c.Flags.ConfigFile
-	} else if f == "example" || f == "---" {
-		// Bubilding a default template.
-		f = c.Flags.ConfigFile
-		c.Config.LogFile = ""
-		c.Config.DebugLog = ""
-		c.Config.HTTPLog = ""
-		c.Config.FileMode = logs.FileMode(rotatorr.FileMode)
-		c.Config.Debug = false
-		c.Config.Snapshot.Interval.Duration = mnd.HalfHour
-		configfile.ForceAllTmpl = true
-	}
-
-	c.Printf("%s", msg)
-
-	f, err := c.Config.Write(f)
-	if err != nil { // f purposely shadowed.
-		return fmt.Errorf("writing config: %s: %w", f, err)
-	}
-
-	c.Print("Wrote Config File:", f)
-
-	return nil
-}
-
-func (c *Client) startupMessage(msg []string) {
-	if ui.HasGUI() {
-		// Setting AppName forces log files (even if not configured).
-		// Used for GUI apps that have no console output.
-		c.Config.LogConfig.AppName = c.Flags.Name()
-	}
-
-	c.Logger.SetupLogging(c.Config.LogConfig)
-	c.Printf("%s v%s-%s Starting! [PID: %v]", c.Flags.Name(), version.Version, version.Revision, os.Getpid())
-
-	for _, m := range msg {
-		c.Printf("==> %s", m)
-	}
-}
-
+// start runs from Start() after the configuration is loaded.
 func (c *Client) start() error {
-	if c.Flags.Updated {
-		c.printUpdateMessage()
-	}
-
-	if c.Flags.TestSnaps {
-		c.checkPlex()
-		c.Config.Snapshot.Validate()
-
-		if c.Flags.SendSnaps {
-			c.configureNotifiarr()
-			c.notify.Start(c.Config.Mode)
-			c.Printf("[user requested] Snapshot Data:\n%s", c.sendSystemSnapshot(c.notify.URL))
-		} else {
-			c.logSnaps()
-		}
-
-		return nil
-	}
-
-	if c.Config.APIKey == "" {
-		return fmt.Errorf("%w %s_API_KEY", ErrNilAPIKey, c.Flags.EnvPrefix)
-	}
-
-	c.configureServices(!c.Flags.CFsync) // do not collect plex info if cfsync is active.
-
-	if c.Flags.CFsync {
-		c.Printf("==> Flag Requested: Syncing Custom Formats and Release Profiles (then exiting)")
-		// c.notify.SendFinishedQueueItems(c.notify.BaseURL)
-		c.notify.SyncCF(true)
-
-		return nil
-	}
-
-	if err := c.Config.Services.Start(c.Config.Service); err != nil {
+	if err := c.configureServices(); err != nil {
 		return fmt.Errorf("service checks: %w", err)
-	}
-
-	if ci, err := c.notify.GetClientInfo(); err != nil {
+	} else if ci, err := c.notifiarr.GetClientInfo(); err != nil {
 		c.Printf("==> [WARNING] API Key may be invalid: %v: %s", err, ci)
 	} else if ci != nil {
 		c.Printf("==> %s", ci)
 	}
 
-	return c.run()
-}
-
-func printProcessList() error {
-	pslist, err := services.GetAllProcesses()
-	if err != nil {
-		return fmt.Errorf("unable to get processes: %w", err)
+	if c.newCon {
+		_, _ = c.Config.Write(c.Flags.ConfigFile)
+		_ = ui.OpenFile(c.Flags.ConfigFile)
+		_, _ = ui.Warning(mnd.Title, "A new configuration file was created @ "+
+			c.Flags.ConfigFile+" - it should open in a text editor. "+
+			"Please edit the file and reload this application using the tray menu.")
+	} else if c.Config.AutoUpdate != "" {
+		go c.AutoWatchUpdate() // do not run updater if there's a brand new config file.
 	}
 
-	for _, p := range pslist {
-		if runtime.GOOS == "freebsd" {
-			fmt.Printf("[%-5d] %s\n", p.PID, p.CmdLine)
-			continue
-		}
-
-		t := "unknown"
-		if !p.Created.IsZero() {
-			t = time.Since(p.Created).Round(time.Second).String()
-		}
-
-		fmt.Printf("[%-5d] %-11s: %s\n", p.PID, t, p.CmdLine)
+	if ui.HasGUI() {
+		// This starts the web server and calls os.Exit() when done.
+		c.startTray()
 	}
 
-	return nil
+	return c.Exit()
 }
 
-// configureServices is called on startup and on reload.
-func (c *Client) configureServices(getPlexInfo bool) {
-	c.checkPlex() // This runs plex.Validate().
+// configureServices is called on startup and on reload, so be careful what goes in here.
+func (c *Client) configureServices() error {
+	if c.Config.Plex.Configured() {
+		_ = c.Config.Plex.Validate()
 
-	if getPlexInfo && c.Config.Plex.Configured() {
 		if info, err := c.Config.Plex.GetInfo(); err != nil {
 			c.Config.Plex.Name = ""
 			c.Errorf("=> Getting Plex Media Server info (check url and token): %v", err)
@@ -278,59 +184,85 @@ func (c *Client) configureServices(getPlexInfo bool) {
 		}
 	}
 
-	c.configureNotifiarr()
 	c.Config.Snapshot.Validate()
 	c.PrintStartupInfo()
-	c.notify.Start(c.Config.Mode)
+	c.notifiarr.Start(c.Config.Mode)
 
-	c.Config.Services.Logger = c.Logger
-	c.Config.Services.Apps = c.Config.Apps
-	c.Config.Services.Notify = c.notify
+	if err := c.Config.Services.Start(c.Config.Service); err != nil {
+		return fmt.Errorf("service checks: %w", err)
+	}
+
+	return nil
 }
 
-func (c *Client) configureNotifiarr() {
-	c.notify = &notifiarr.Config{
-		Apps:    c.Config.Apps,
-		Plex:    c.Config.Plex,
-		Snap:    c.Config.Snapshot,
-		Logger:  c.Logger,
-		URL:     notifiarr.ProdURL,
-		Timeout: c.Config.Timeout.Duration,
-		DashDur: c.Config.SendDash.Duration,
-	}
-}
+// Exit stops the web server and logs our exit messages. Start() calls this.
+func (c *Client) Exit() error {
+	c.StartWebServer()
+	c.setSignals()
 
-// run turns on the auto updater if enabled, and starts the web server, and system tray icon.
-func (c *Client) run() error {
-	signal.Notify(c.sigkil, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT, syscall.SIGHUP)
-	c.setReloadSignals()
-
-	if c.newCon {
-		_, _ = c.Config.Write(c.Flags.ConfigFile)
-		_ = ui.OpenFile(c.Flags.ConfigFile)
-		_, _ = ui.Warning(mnd.Title, "A new configuration file was created @ "+
-			c.Flags.ConfigFile+" - it should open in a text editor. "+
-			"Please edit the file and reload this application using the tray menu.")
-	}
-
-	if c.Config.AutoUpdate != "" {
-		go c.AutoWatchUpdate()
-	}
-
-	switch ui.HasGUI() {
-	case true:
-		c.startTray() // This starts the web server.
-		return nil    // startTray() calls os.Exit()
-	default:
-		c.StartWebServer()
-		return c.Exit()
+	// For non-GUI systems, this is where the main go routine stops (and waits).
+	for {
+		select {
+		case sigc := <-c.sigkil:
+			c.Printf("[%s] Need help? %s\n=====> Exiting! Caught Signal: %v", c.Flags.Name(), mnd.HelpLink, sigc)
+			return c.exit()
+		case sigc := <-c.sighup:
+			if err := c.checkReloadSignal(sigc); err != nil {
+				return err // reloadConfiguration()
+			}
+		}
 	}
 }
 
-// starts plex if it's configured.
-func (c *Client) checkPlex() {
-	if c.Config.Plex.Configured() {
-		// Validate only returns an error if Configured == false.
-		_ = c.Config.Plex.Validate()
+func (c *Client) exit() error {
+	if c.server == nil {
+		return nil
 	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), c.Config.Timeout.Duration)
+	defer cancel()
+
+	if err := c.server.Shutdown(ctx); err != nil {
+		return fmt.Errorf("server shutdown failed: %w", err)
+	}
+
+	c.server = nil
+
+	return nil
+}
+
+// reloadConfiguration is called from a menu tray item or when a HUP signal is received.
+// Re-reads the configuration file and stops/starts all the internal routines.
+// Also closes and re-opens all log files. Any errors cause the application to exit.
+func (c *Client) reloadConfiguration(source string) error {
+	c.Print("==> Reloading Configuration: " + source)
+	c.notifiarr.Stop()
+	c.Config.Services.Stop()
+
+	err := c.StopWebServer()
+	if err != nil && !errors.Is(err, ErrNoServer) {
+		return fmt.Errorf("stoping web server: %w", err)
+	} else if err == nil {
+		defer c.StartWebServer()
+	}
+
+	c.notifiarr, err = c.Config.Get(c.Flags.ConfigFile, c.Flags.EnvPrefix, c.Logger)
+	if err != nil {
+		return fmt.Errorf("getting configuration: %w", err)
+	}
+
+	if errs := c.Logger.Close(); len(errs) > 0 {
+		return fmt.Errorf("closing logger(s): %w", errs[0])
+	}
+
+	c.Logger.SetupLogging(c.Config.LogConfig)
+
+	if err := c.configureServices(); err != nil {
+		return err
+	}
+
+	c.Print("==> Configuration Reloaded! Config File:", c.Flags.ConfigFile)
+	_, _ = ui.Info(mnd.Title, "Configuration Reloaded!") //nolint:wsl
+
+	return nil
 }
