+++ pkg/client/tray.go
@@ -19,44 +19,42 @@ import (
 
 // startTray Run()s readyTray to bring up the web server and the GUI app.
 func (c *Client) startTray() {
-	systray.Run(c.readyTray, c.exitTray)
-}
+	systray.Run(func() {
+		defer os.Exit(0)
+		defer c.CapturePanic()
 
-func (c *Client) exitTray() {
-	c.sigkil = nil
-
-	if err := c.Exit(); err != nil {
-		c.Errorf("Shutting down web server: %v", err)
-		os.Exit(1) // web server problem
-	}
-	// because systray wants to control the exit code? no..
-	os.Exit(0)
-}
-
-// readyTray creates the system tray/menu bar app items, and starts the web server.
-func (c *Client) readyTray() {
-	b, err := bindata.Asset(ui.SystrayIcon)
-	if err == nil {
+		b, _ := bindata.Asset(ui.SystrayIcon)
 		systray.SetTemplateIcon(b, b)
-	} else {
-		c.Errorf("Reading Icon: %v", err)
-		systray.SetTitle("DNC")
+		systray.SetTooltip(c.Flags.Name() + " v" + version.Version)
+		c.makeChannels() // make these before starting the web server.
+		c.makeMoreChannels()
+		c.setupChannels(c.watchKillerChannels, c.watchNotifiarrMenu,
+			c.watchLogsChannels, c.watchConfigChannels, c.watchGuiChannels)
+
+		// This starts the web server, and waits for reload/exit signals.
+		if err := c.Exit(); err != nil {
+			c.Errorf("Server: %v", err)
+			os.Exit(1) // web server problem
+		}
+	}, func() {
+		// This code only fires from menu->quit.
+		if err := c.exit(); err != nil {
+			c.Errorf("Server: %v", err)
+			os.Exit(1) // web server problem
+		}
+		// because systray wants to control the exit code? no..
+		os.Exit(0)
+	})
+}
+
+// setupChannels runs the channel watcher loops in go routines with a panic catcher.
+func (c *Client) setupChannels(funcs ...func()) {
+	for _, f := range funcs {
+		go func(f func()) {
+			defer c.CapturePanic()
+			f()
+		}(f)
 	}
-
-	systray.SetTooltip(c.Flags.Name() + " v" + version.Version)
-
-	c.makeChannels() // make these before starting the web server.
-	c.makeMoreChannels()
-	c.menu["info"].Disable()
-	c.menu["dninfo"].Hide()
-	c.menu["alert"].Hide() // currently unused.
-
-	go c.watchKillerChannels()
-	go c.watchNotifiarrMenu()
-	go c.watchLogsChannels()
-	go c.watchConfigChannels()
-	c.StartWebServer()
-	c.watchGuiChannels()
 }
 
 func (c *Client) makeChannels() {
@@ -73,6 +71,7 @@ func (c *Client) makeChannels() {
 	link := systray.AddMenuItem("Links", "external resources")
 	c.menu["link"] = ui.WrapMenu(link)
 	c.menu["info"] = ui.WrapMenu(link.AddSubMenuItem(c.Flags.Name(), version.Print(c.Flags.Name())))
+	c.menu["info"].Disable()
 	c.menu["hp"] = ui.WrapMenu(link.AddSubMenuItem("Notifiarr.com", "open Notifiarr.com"))
 	c.menu["wiki"] = ui.WrapMenu(link.AddSubMenuItem("Notifiarr Wiki", "open Notifiarr wiki"))
 	c.menu["disc1"] = ui.WrapMenu(link.AddSubMenuItem("Notifiarr Discord", "open Notifiarr discord server"))
@@ -91,6 +90,7 @@ func (c *Client) makeChannels() {
 	}
 }
 
+// makeMoreChannels makes the Notifiarr menu and Debug menu items.
 //nolint:lll
 func (c *Client) makeMoreChannels() {
 	data := systray.AddMenuItem("Notifiarr", "plex sessions, system snapshots, service checks")
@@ -115,7 +115,7 @@ func (c *Client) makeMoreChannels() {
 	c.menu["debug_logs"] = ui.WrapMenu(debug.AddSubMenuItem("View Log", "view the Debug log"))
 	// debug.AddSeparator() // not exist: https://github.com/getlantern/systray/issues/170
 	ui.WrapMenu(debug.AddSubMenuItem("__________", "")).Disable() // fake separator.
-	c.menu["debug_panic"] = ui.WrapMenu(debug.AddSubMenuItem("Panic", "cause an application panic"))
+	c.menu["debug_panic"] = ui.WrapMenu(debug.AddSubMenuItem("Panic", "cause an application panic (crash)"))
 
 	if c.Config.DebugLog == "" {
 		c.menu["debug_logs"].Hide()
@@ -131,10 +131,11 @@ func (c *Client) makeMoreChannels() {
 		c.menu["debug"].Hide()
 	}
 
-	// These start hidden.
 	c.menu["update"] = ui.WrapMenu(systray.AddMenuItem("Update", "Check GitHub for Update"))
 	c.menu["dninfo"] = ui.WrapMenu(systray.AddMenuItem("Info!", "info from Notifiarr.com"))
+	c.menu["dninfo"].Hide()
 	c.menu["alert"] = ui.WrapMenu(systray.AddMenuItem("Alert!", "alert from Notifiarr.com"))
+	c.menu["alert"].Hide() // currently unused.
 
 	c.menu["exit"] = ui.WrapMenu(systray.AddMenuItem("Quit", "Exit "+c.Flags.Name()))
 }
@@ -144,12 +145,6 @@ func (c *Client) watchKillerChannels() {
 
 	for {
 		select {
-		case sigc := <-c.sighup:
-			c.Printf("Caught Signal: %v (reloading configuration)", sigc)
-			c.reloadConfiguration("caught signal " + sigc.String())
-		case sigc := <-c.sigkil:
-			c.Errorf("Need help? %s\n=====> Exiting! Caught Signal: %v", mnd.HelpLink, sigc)
-			return
 		case <-c.menu["exit"].Clicked():
 			c.Errorf("Need help? %s\n=====> Exiting! User Requested", mnd.HelpLink)
 			return
@@ -161,6 +156,11 @@ func (c *Client) watchKillerChannels() {
 		case <-c.menu["debug_logs"].Clicked():
 			c.Print("User Viewing Debug File:", c.Config.DebugLog)
 			_ = ui.OpenLog(c.Config.DebugLog)
+		case <-c.menu["load"].Clicked():
+			if err := c.reloadConfiguration("User Requested"); err != nil {
+				c.Errorf("Need help? %s\n=====> Exiting! Reloading Configuration: %v", mnd.HelpLink, err)
+				os.Exit(1) //nolint:gocritic // exit now since config is bad and everything is disabled.
+			}
 		}
 	}
 }
@@ -194,8 +194,6 @@ func (c *Client) watchConfigChannels() {
 		case <-c.menu["edit"].Clicked():
 			c.Print("User Editing Config File:", c.Flags.ConfigFile)
 			ui.OpenFile(c.Flags.ConfigFile)
-		case <-c.menu["load"].Clicked():
-			c.reloadConfiguration("UI requested")
 		case <-c.menu["write"].Clicked():
 			c.writeConfigFile()
 		case <-c.menu["key"].Clicked():
@@ -233,7 +231,7 @@ func (c *Client) watchNotifiarrMenu() { //nolint:cyclop
 		select {
 		case <-c.menu["sync_cf"].Clicked():
 			c.Printf("[user requested] Triggering Custom Formats and Quality Profiles Sync for Radarr and Sonarr.")
-			c.notify.SyncCF(true)
+			c.notifiarr.Trigger.SyncCF(false)
 		case <-c.menu["snap_log"].Clicked():
 			c.logSnaps()
 		case <-c.menu["svcs_log"].Clicked():
@@ -270,16 +268,16 @@ func (c *Client) watchNotifiarrMenu() { //nolint:cyclop
 		case <-c.menu["snap_dev"].Clicked():
 			c.sendSystemSnapshot(notifiarr.DevURL)
 		case <-c.menu["app_ques"].Clicked():
-			c.notify.SendFinishedQueueItems(notifiarr.BaseURL)
+			c.notifiarr.Trigger.SendFinishedQueueItems(notifiarr.BaseURL)
 		case <-c.menu["app_ques_dev"].Clicked():
-			c.notify.SendFinishedQueueItems(notifiarr.DevBaseURL)
+			c.notifiarr.Trigger.SendFinishedQueueItems(notifiarr.DevBaseURL)
 		case <-c.menu["plex_prod"].Clicked():
 			c.sendPlexSessions(notifiarr.ProdURL)
 		case <-c.menu["snap_prod"].Clicked():
 			c.sendSystemSnapshot(notifiarr.ProdURL)
 		case <-c.menu["send_dash"].Clicked():
 			c.Print("User Requested State Collection for Dashboard")
-			c.notify.GetState()
+			c.notifiarr.Trigger.GetState()
 		}
 	}
 }
