--- pkg/client/tray_other.go
@@ -1,5 +1,0 @@
-// +build !windows,!darwin
-
-package client
-
-func (c *Client) startTray() {}
