+++ pkg/logs/logs.go
@@ -12,8 +12,10 @@ import (
 	"log"
 	"os"
 	"path/filepath"
+	"runtime/debug"
 
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
+	"github.com/Notifiarr/notifiarr/pkg/ui"
 	homedir "github.com/mitchellh/go-homedir"
 	"golift.io/rotatorr"
 	"golift.io/rotatorr/timerotator"
@@ -150,6 +152,16 @@ func (l *Logger) Close() (errors []error) {
 	}
 
 	return errors
+}
+
+// CapturePanic can be defered in any go routine to log any panic that occurs.
+func (l *Logger) CapturePanic() {
+	ui.ShowConsoleWindow()
+
+	if r := recover(); r != nil {
+		_ = l.ErrorLog.Output(callDepth,
+			fmt.Sprintf("Go Panic! %s\n%v\n%s", mnd.BugIssue, r, string(debug.Stack())))
+	}
 }
 
 // Debug writes log lines... to stdout and/or a file.
