+++ pkg/notifiarr/plexcron.go
@@ -23,8 +23,8 @@ const (
 )
 
 // SendPlexSessions sends plex sessions in a go routine through a channel.
-func (c *Config) SendPlexSessions(source string) {
-	c.plexNow <- source
+func (t *triggers) SendPlexSessions(source string) {
+	t.plex <- source
 }
 
 // sendPlexSessions is fired by a timer if plex monitoring is enabled.
