+++ pkg/notifiarr/state.go
@@ -80,8 +80,8 @@ type States struct {
 	Deluge  []*State `json:"deluge"`
 }
 
-func (c *Config) GetState() {
-	c.stateNow <- struct{}{}
+func (t *triggers) GetState() {
+	t.state <- struct{}{}
 }
 
 func (c *Config) getState() {
@@ -112,26 +112,32 @@ func (c *Config) getStates() *States {
 	wg.Add(6) //nolint:gomnd // we are polling 6 apps.
 
 	go func() {
+		defer c.CapturePanic()
 		s.Deluge = c.getDelugeStates()
 		wg.Done() //nolint:wsl
 	}()
 	go func() {
+		defer c.CapturePanic()
 		s.Lidarr = c.getLidarrStates()
 		wg.Done() //nolint:wsl
 	}()
 	go func() {
+		defer c.CapturePanic()
 		s.Qbit = c.getQbitStates()
 		wg.Done() //nolint:wsl
 	}()
 	go func() {
+		defer c.CapturePanic()
 		s.Radarr = c.getRadarrStates()
 		wg.Done() //nolint:wsl
 	}()
 	go func() {
+		defer c.CapturePanic()
 		s.Readarr = c.getReadarrStates()
 		wg.Done() //nolint:wsl
 	}()
 	go func() {
+		defer c.CapturePanic()
 		s.Sonarr = c.getSonarrStates()
 		wg.Done() //nolint:wsl
 	}()
