+++ pkg/ui/ui_darwin.go
@@ -19,6 +19,9 @@ func HasGUI() bool {
 // HideConsoleWindow doesn't work on maacOS.
 func HideConsoleWindow() {}
 
+// ShowConsoleWindow does nothing on OSes besides Windows.
+func ShowConsoleWindow() {}
+
 // StartCmd starts a command.
 func StartCmd(c string, v ...string) error {
 	cmd := exec.Command(c, v...)
