+++ pkg/ui/ui_other.go
@@ -20,6 +20,9 @@ func HasGUI() bool {
 // HideConsoleWindow doesn't work on most OSes.
 func HideConsoleWindow() {}
 
+// ShowConsoleWindow does nothing on OSes besides Windows.
+func ShowConsoleWindow() {}
+
 // StartCmd starts a command.
 func StartCmd(c string, v ...string) error {
 	cmd := exec.Command(c, v...)
